Add initial implementation for improved dev scripts experience
