+++ .eslintrc
@@ -1,5 +1,12 @@
 {
+  "root": true,
+
   "extends": ["@ikscodes/eslint-config"],
+
+  "parserOptions": {
+    "project": ["./packages/tsconfig.json", "./scripts/tsconfig.json"]
+  },
+
   "rules": {
     "import/extensions": 0,
     "no-alert": 0,
@@ -14,15 +21,11 @@
     "@typescript-eslint/no-unsafe-call": 0,
     "@typescript-eslint/ban-types": 0,
     "@typescript-eslint/explicit-module-boundary-types": 0,
+    "@typescript-eslint/restrict-template-expressions": 0,
     "class-methods-use-this": 0,
     "import/no-extraneous-dependencies": 0
   },
-  "parserOptions": {
-    "project": [
-      "./tsconfig.json",
-      "./test/tsconfig.json"
-    ]
-  },
+
   "settings": {
     "import/resolver": {
       "typescript": {
