+++ packages/provider/.eslintignore
@@ -1,4 +1,2 @@
 /node_modules
 /dist
-types.js
-types.d.ts
