+++ packages/provider/tsconfig.module.json
@@ -2,8 +2,9 @@
   "extends": "../tsconfig.settings.json",
 
   "compilerOptions": {
+    "module": "esnext",
     "rootDir": "./src",
-    "outDir": "./dist/commonjs"
+    "outDir": "./dist/module"
   },
 
   "include": [
@@ -12,6 +13,6 @@
   ],
 
   "references": [
-    { "path": "../types" }
+    { "path": "../types" },
   ]
 }
