--- packages/react-native/.babelrc
@@ -1,7 +1,0 @@
-{
-	"plugins": [
-		[
-			"@babel/plugin-proposal-optional-chaining"
-		]
-	]
-}
