+++ packages/react-native/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extends: ['../../.eslintrc'],
+  parserOptions: {
+    project: ['./tsconfig.json', './test/tsconfig.json'],
+    tsconfigRootDir: __dirname,
+  },
+}
