+++ packages/react-native/tsconfig.json
@@ -1,13 +1,18 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.settings.json",
+
   "compilerOptions": {
-    "lib": ["esnext", "dom"],
-    "module": "esnext",
     "rootDir": "./src",
-    "outDir": "./dist"
+    "outDir": "./dist/commonjs"
   },
+
   "include": [
     "./src/**/*.ts",
     "./src/**/*.tsx",
+  ],
+
+  "references": [
+    { "path": "../types" },
+    { "path": "../provider" },
   ]
 }
