+++ packages/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "files": [],
+
+  "references": [
+    { "path": "./provider/tsconfig.json" },
+    { "path": "./provider/tsconfig.module.json" },
+
+    { "path": "./react-native/tsconfig.json" },
+    { "path": "./react-native/tsconfig.module.json" },
+
+    { "path": "./types/tsconfig.json" },
+    { "path": "./types/tsconfig.module.json" },
+
+    { "path": "./web/tsconfig.json" },
+    { "path": "./web/tsconfig.module.json" },
+  ]
+}
