+++ packages/types/tsconfig.json
@@ -1,12 +1,12 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.settings.json",
+
   "compilerOptions": {
-    "lib": ["esnext", "dom"],
     "rootDir": "./src",
-    "outDir": "./dist"
+    "outDir": "./dist/commonjs"
   },
+
   "include": [
     "./src/**/*.ts",
-    "./src/**/*.tsx",
-  ]
+  ],
 }
