+++ packages/types/tsconfig.module.json
@@ -2,16 +2,12 @@
   "extends": "../tsconfig.settings.json",
 
   "compilerOptions": {
+    "module": "esnext",
     "rootDir": "./src",
-    "outDir": "./dist/commonjs"
+    "outDir": "./dist/module"
   },
 
   "include": [
     "./src/**/*.ts",
-    "./src/**/*.js",
   ],
-
-  "references": [
-    { "path": "../types" }
-  ]
 }
