+++ packages/web/tsconfig.cdn.json
@@ -1,7 +1,19 @@
 {
-  "extends": "./tsconfig.json",
+  "extends": "../tsconfig.settings.json",
+
   "compilerOptions": {
+    "module": "esnext",
     "declaration": false,
     "sourceMap": false,
-  }
+    "composite": false,
+  },
+
+  "include": [
+    "./src/**/*.ts",
+  ],
+
+  "references": [
+    { "path": "../types" },
+    { "path": "../provider" },
+  ]
 }
