+++ packages/web/tsconfig.json
@@ -1,11 +1,17 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.settings.json",
+
   "compilerOptions": {
-    "lib": ["esnext", "dom"],
     "rootDir": "./src",
-    "outDir": "./dist"
+    "outDir": "./dist/commonjs"
   },
+
   "include": [
     "./src/**/*.ts",
+  ],
+
+  "references": [
+    { "path": "../types" },
+    { "path": "../provider" },
   ]
 }
