+++ scripts/clean.sh
@@ -1,9 +1,5 @@
 #!/usr/bin/env bash
 
-echo
-echo "Cleaning up generated files..."
-echo
-
 CLEAN_DIST=true
 CLEAN_CACHE=false
 CLEAN_TEST_ARTIFACTS=false
@@ -33,14 +29,14 @@ while test $# -gt 0; do
   esac
 done
 
-if [ $PKG ] ; then
-  if [ $CLEAN_DIST = true ]; then lerna exec --scope $PKG -- rimraf dist ; fi
-  if [ $CLEAN_CACHE = true ]; then lerna exec --scope $PKG -- rimraf node_modules/.cache ; fi
-  if [ $CLEAN_TEST_ARTIFACTS = true ]; then lerna exec --scope $PKG -- rimraf coverage && rimraf .nyc_output ; fi
-  if [ $CLEAN_NODE_MODULES = true ]; then lerna exec --scope $PKG -- rimraf node_modules ; fi
-else
-  if [ $CLEAN_DIST = true ]; then lerna exec -- rimraf dist ; fi
-  if [ $CLEAN_CACHE = true ]; then lerna exec -- rimraf node_modules/.cache ; fi
-  if [ $CLEAN_TEST_ARTIFACTS = true ]; then lerna exec -- rimraf coverage && rimraf .nyc_output ; fi
-  if [ $CLEAN_NODE_MODULES = true ]; then lerna exec -- rimraf node_modules ; fi
-fi
+msg() {
+  echo
+  echo "+------------------------------------------------------------------------------+"
+  echo "  Cleaning $1..."
+  echo
+}
+
+if [ $CLEAN_DIST = true ]; then msg "build files" && yarn wsrun --parallel -r rimraf dist ; fi
+if [ $CLEAN_CACHE = true ]; then msg "caches" && yarn wsrun --parallel -r rimraf node_modules/.cache ; fi
+if [ $CLEAN_TEST_ARTIFACTS = true ]; then msg "test artifacts" && (yarn wsrun --parallel -r rimraf coverage) && (yarn wsrun --parallel -r rimraf .nyc_output) ; fi
+if [ $CLEAN_NODE_MODULES = true ]; then msg "node_modules" && yarn wsrun --parallel -r rimraf node_modules && rimraf node_modules ; fi
