+++ scripts/lint.sh
@@ -1,11 +1,8 @@
 #!/usr/bin/env bash
 
 echo
-echo "Running the linter."
+echo "+------------------------------------------------------------------------------+"
+echo "  Linting TypeScripts..."
 echo
 
-if [ $PKG ] ; then
-  lerna exec --scope $PKG -- eslint --fix src/**/*.ts
-else
-  lerna exec -- eslint --fix src/**/*.ts
-fi
+yarn wsrun --stages eslint --fix "**/src/**/*.ts"
