+++ scripts/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../packages/tsconfig.settings.json",
+  "compilerOptions": {
+    "noImplicitAny": false,
+    "composite": false,
+  },
+  "include": ["./**/*.ts"]
+}
